split fixtures into data and common, created new fixtures folder and updated unit tests
